Alterada classe Roupa para Produto e toda refatoração

+++ src/main/java/com/estoque/lelu/service/FornecedorService.java
@@ -1,6 +1,7 @@
 package com.estoque.lelu.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -18,12 +19,16 @@ public class FornecedorService {
 		return fornecedorRepository.save(fornecedor);
 	}
 
-	public Fornecedor buscarFornecedorPorId(Long id) {
-		return fornecedorRepository.findById(id).orElse(null);
+	public Optional<Fornecedor> buscarFornecedorPorId(Long id) {
+		return fornecedorRepository.findById(id);
 	}
 
-	public void deletarFornecedor(Long id) {
-		fornecedorRepository.deleteById(id);
+	public boolean deletarFornecedor(Long id) {
+		if (fornecedorRepository.existsById(id)) {
+			fornecedorRepository.deleteById(id);
+			return true;
+		}
+		return false;
 	}
 
 	public List<Fornecedor> listarFornecedores() {
